Add tests for TRUNCATE TABLE
FIXME: A prerequisite for crash-safe TRUNCATE is a crash-safe RENAME TABLE
(with proper rollback).

FIXME: Reduce the redo log volume for undo tablespace truncation,
and remove the workarounds in --suite=innodb_undo. The redo log
format for undo tablespace truncation will be changing anyway,
so we might as well change it further for the initialization of the
rollback segment header pages.